agent-system/requirements.txt
# Google ADK & Gemini
google-genai>=0.1.0
google-generativeai>=0.8.0

# MCP (Model Context Protocol)
mcp>=0.1.0

# OpenTelemetry (Observability)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-jaeger-thrift>=1.20.0
opentelemetry-instrumentation>=0.41b0

# Prometheus & Metrics
prometheus-client>=0.19.0

# MongoDB
pymongo>=4.6.0
motor>=3.3.2

# HTTP & APIs
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0

# OpenAI (for embeddings)
openai>=1.6.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# FastAPI (for agent API server)
fastapi>=0.108.0
uvicorn[standard]>=0.25.0

# Jupyter (for examples)
jupyter>=1.0.0
ipykernel>=6.27.0
